chore(merge-executor): remove merge-response variants with no producer
The executor builds its own merge commit and fast-forwards the target ref, so
`head-moved` and `not-mergeable` — the outcomes only the retired pull-request
merge API could return — had no producer, and the decision-table branches
consuming them were dead. The tests covering those branches passed against a
fake alone, so they asserted nothing about production code.

- Drop both variants and the decision-table branches that consumed them, and
  refresh the comments still describing the PUT path's 405 and expected-head
  compare-and-swap.
- Rewrite the affected tests against outcomes the fast-forward path really
  produces: a refused `beforeOid` CAS named `base-drift` by the read-back, and
  a refused ref update classified by exactly one re-read.
- Inline `fake-pr-surface.ts` into its only importer and delete it, so the
  isolation scan no longer needs an exclusion for a test-only source file.
- Add a closed-set guard asserting every `MergeResponse` variant has a producer
  in `github.ts`, so an orphaned variant fails here rather than lingering.

The remaining deterministic-rejection variants (`forbidden`, `not-found`,
`unprocessable`) gained producers in `restStop` after the brief was written and
are kept, with their existing 403/422 class tests.


X-Anneal-Run: cmtq08tvk0oj7db3eze7czci2
X-Anneal-Step: 1: Implementation
X-Anneal-Provider: claude
